Create balances
In this version, I've removed the balances mapping because it does a similar job to loanAmounts, just more intuitive in this context. I also removed the payable modifier since ether is no longer received directly from the contract. Additionally, I used a new modifier onlyOwner to ensure that only the owner of the contract can perform certain actions